Fixed bug causing error with agencies sometimes being added before their counties were, and handling counties without names, performed light refactoring including adding error handling.
